added: descriptive comments for ModalMolecule.jsx
